Extract searchFromSelection
