+++ src/pages/Content/index.js
@@ -3,34 +3,13 @@ import ReactDOM from 'react-dom'
 import ContentApp from './ContentApp'
 import { QueryClient, QueryClientProvider } from 'react-query'
 
-import store from '../../lib/store'
-
-document.addEventListener('dblclick', async () => {
-  searchFromSelection()
-})
+import { searchFromSelection } from '../../lib/store'
 
 chrome.runtime.onMessage.addListener(function ({ type }, sender, sendResponse) {
   if (type === 'mojidict:searchSelection') {
     searchFromSelection()
   }
 })
-
-function searchFromSelection() {
-  const selection = window.getSelection()
-  const selectionText = selection?.toString().trim()
-  const range = selection?.getRangeAt(0)
-
-  if (!selectionText.length || !range) {
-    return
-  }
-
-  store.setState((state) => ({
-    ...state,
-    searchKeyword: selectionText,
-    selectionRect: range.getBoundingClientRect(),
-    showCard: true,
-  }))
-}
 
 function getWordCardContainer() {
   return document.querySelector('.mojidict-helper-card-container')
